+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Playlist Tools\Playlist manipulation: new submenu with playlists sorting features.
 - Playlist Tools\Playlist manipulation: new submenu with playlists creation features.
 - Playlist Tools\Playlist manipulation: new submenu with playlists creation features.
+- Playlist Tools\Playlist manipulation: new submenu to perform a query search on specific playlists (which also works with dynamic queries based on current selection).
 - Quicksearch: added new setting for max tracks from selection processed by dynamic queries. Previously value was hardcoded (default 1000).
 ### Changed
 - Playlist Tools\Playlist manipulation: menu cleanup and minor reordering.